feat: 분리된 SSR+CSR 구조로 RootFlow 구현
- app/page.tsx를 SSR 서버 컴포넌트로 변경하여 RootFlow를 포함
- RootFlow.tsx를 CSR island로 분리해 Splash → Onboarding → AuthLanding 단계 제어
